+++ app/controllers/admin/claims_controller.rb
@@ -18,6 +18,7 @@ class Admin::ClaimsController < Admin::BaseAdminController
     @claim = Claim.find(params[:id])
     @check = @claim.check || Check.new
     @matching_claims = Claim::MatchingAttributeFinder.new(@claim).matching_claims
+    @claims_preventing_payment = Claim::ClaimsPreventingPaymentFinder.new(@claim).claims_preventing_payment
   end
 
   def search